unsigned long -> size_t in gauss

+++ linbox/algorithms/gauss/gauss-elim-gf2.inl
@@ -32,10 +32,10 @@ namespace LinBox
 {
 	template <class Vector> inline void
 	GaussDomain<GF2>::permuteBinary (Vector              &lignecourante,
-					 const unsigned long &indcol,
+					 const size_t &indcol,
 					 const long &indpermut) const
 	{
-		const unsigned long k = indcol - 1;
+		const size_t k = indcol - 1;
 
 #if 0
 		std::cerr << "B PERMUTE: " << indpermut << " <--> " << k << " of  [";
@@ -114,16 +114,16 @@ namespace LinBox
 	GaussDomain<GF2>::eliminateBinary (bool             &headpivot,
 					   Vector              &lignecourante,
 					   const Vector        &lignepivot,
-					   const unsigned long indcol,
+					   const size_t indcol,
 					   const long indpermut,
-					   const unsigned long npiv,
+					   const size_t npiv,
 					   D                   &columns) const
 	{
 
 		typedef typename Vector::value_type E;
 
-		unsigned long k = indcol - 1;
-		unsigned long nj = lignecourante.size () ;
+		size_t k = indcol - 1;
+		size_t nj = lignecourante.size () ;
 #if 0
 		std::cerr << "BEGIN ELIMINATE, k: " << k << ", nj: " << nj << ", indpermut: " << indpermut << ", indcol: " << indcol << std::endl;
 		std::cerr << "lignepivot: [";
@@ -139,7 +139,7 @@ namespace LinBox
 		std::cerr << ']' << std::endl;
 #endif
 		if (nj > 0) {
-			unsigned long j_head = 0;
+			size_t j_head = 0;
 
 			for (; j_head < nj; ++j_head) {
 				if (static_cast<long>(lignecourante[(size_t)j_head]) >= indpermut) break;
@@ -174,8 +174,8 @@ namespace LinBox
 					// construit : <-- j
 					// courante  : <-- m
 					// pivot     : <-- l
-					unsigned long j = 0;
-					unsigned long m = j_head + 1;
+					size_t j = 0;
+					size_t m = j_head + 1;
 
 					// A[i,k] <-- - A[i,k] / A[k,k]
 					headpivot = true;
@@ -188,14 +188,14 @@ namespace LinBox
 					}
 
 
-					unsigned long l = 0;
+					size_t l = 0;
 
 					for (; l < npiv; ++l)
 						if (lignepivot[l] > k) break;
 
 					// for all j such that (j>k) and A[k,j]!=0
 					while (l < npiv) {
-					unsigned long j_piv;
+					size_t j_piv;
 						j_piv = lignepivot[l];
 
 						// if A[k,j]=0, then A[i,j] <-- A[i,j]
@@ -236,7 +236,7 @@ namespace LinBox
 #endif
 					if (indpermut != static_cast<long>(k)) {
 						if (j_head>0) {
-							unsigned long l = 0;
+							size_t l = 0;
 
 							for (; l < nj; ++l)
 								if (lignecourante[(size_t)l] >= k) break;
@@ -248,7 +248,7 @@ namespace LinBox
 								++columns[(size_t)indpermut];
 								tmp = (E)indpermut;
 
-								unsigned long bjh = j_head-1;
+								size_t bjh = j_head-1;
 								for (; l < bjh; ++l)
 									lignecourante[(size_t)l] = lignecourante[(size_t)l + 1];
 
@@ -271,7 +271,7 @@ namespace LinBox
 				std::cerr << "----------------------------------------------------------" << std::endl;
 #endif
 				if (indpermut != static_cast<long>(k)) {
-					unsigned long l = 0;
+					size_t l = 0;
 
 					for (; l < nj; ++l)
 						if (lignecourante[(size_t)l] >= k) break;
@@ -283,7 +283,7 @@ namespace LinBox
 						++columns[(size_t)indpermut];
 						tmp = (E) indpermut;
 
-						unsigned long bjh = nj - 1;
+						size_t bjh = nj - 1;
 						for (; l < bjh; ++l)
 							lignecourante[(size_t)l] = lignecourante[(size_t)l + 1];
